Create an ASP.NET Core Web API
An ASP.NET Core Web API to create a RESTful web API using controllers that act as intermediaries for files managements, validations and uploads using the referenced LibServices project.

This API will be used to connect different frameworks such as Blazor Webassembly, Angular, Php to the backend libraries and/or the LibServices project.