Configure CORS for API NetCore and Blazor WebAssembly -> Ok
In Blazor WebAssembly you must enter the API domain because that is where Blazor WASM will go for requests.
In the API you must enter the Blazor WebAssembly domain and ALSO the API domain so that both can have access to the endpoints.

Improve user experience on form success/error messages