Create home and upload-file-byte-array. Assign routes for navigation
Creating the home and upload-file-byte-array components.
Registering the components in app.modules.
Setting the app component as home.
Moving the html/css content from app to home.
Creating the button that takes us to the upload-file-byte-array html.
Assign routes for navigation in app-routing.
Formatting the code to accept 'eslint' (prettier/prettier).